backend/requirements.txt
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.4
bcrypt==4.3.0
click==8.3.0
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.121.2
fastapi-mail==1.4.1
greenlet==3.2.4
jinja2==3.1.4
h11==0.16.0
httpx==0.27.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.2.1
python-multipart==0.0.20
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
loguru==0.7.2
